requirements.txt
python-telegram-bot==13.15
vk-api==11.9.9
environs==9.5.0
redis==4.6.0
python-Levenshtein==0.21.1
fuzzywuzzy==0.18.0